requirements.txt
django
djangorestframework-simplejwt
pip install djangorestframework